Update TODOs, Move old components to Deliver component.
