+++ checkout-ui-custom/checkout6-custom.js
@@ -126,7 +126,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \**************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _utils_const__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/const */ \"./src/utils/const.js\");\n/* harmony import */ var _MixedProducts__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../MixedProducts */ \"./src/partials/MixedProducts.js\");\n/* harmony import */ var _TVorRICAMsg__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../TVorRICAMsg */ \"./src/partials/TVorRICAMsg.js\");\n/* harmony import */ var _Addresses__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./Addresses */ \"./src/partials/Deliver/Addresses.js\");\n/* harmony import */ var _AddressForm__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./AddressForm */ \"./src/partials/Deliver/AddressForm.js\");\n/* harmony import */ var _AddressSearch__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./AddressSearch */ \"./src/partials/Deliver/AddressSearch.js\");\n/* harmony import */ var _DeliveryOptions__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./DeliveryOptions */ \"./src/partials/Deliver/DeliveryOptions.js\");\n\n\n\n\n\n\n\n\nconst DeliverContainer = ({\n  hasFurn\n}) => `\n<div id=\"bash--shipping-messages\">\n${(0,_TVorRICAMsg__WEBPACK_IMPORTED_MODULE_2__[\"default\"])()}\n${(0,_MixedProducts__WEBPACK_IMPORTED_MODULE_1__[\"default\"])()}\n</div>\n  <div class=\"bash--delivery-container\" id=\"bash--delivery-container\" data-view=\"select-address\">\n  \n   <form id=\"bash--delivery-form\" name=\"bash--delivery-form\" method=\"post\">\n\n    <section class=\"bash--delivery-view\" data-section=\"select-address\">\n    <div class=\"bash--heading\">\n        <h2>Delivery address</h2>\n        <a href=\"#\" data-view=\"address-search\">Add address</a>\n      </div>\n      ${(0,_Addresses__WEBPACK_IMPORTED_MODULE_3__[\"default\"])()}\n    </section>\n\n    <section id=\"bash-delivery-options\" class=\"shipping-method bash--delivery-view\" data-section=\"select-address\">\n      <hr>\n      <div class=\"bash--heading sub-heading\">\n        <h3>Delivery method</h3>\n        ${hasFurn ? _utils_const__WEBPACK_IMPORTED_MODULE_0__.FURNITURE_FEE_LINK : ''}\n      </div>\n      ${(0,_DeliveryOptions__WEBPACK_IMPORTED_MODULE_6__[\"default\"])()}\n      <button \n        class=\"submit btn-go-to-payment btn btn-large btn-success\"\n        id=\"btn-save-delivery\" \n        type=\"submit\">\n          Go to payment\n      </button>\n    </section>\n   </form>\n\n    <section class=\"bash--delivery-view\" data-section=\"address-search\">\n      <div class=\"bash--heading\">\n        <h3>Add a new delivery address</h3>\n        <a href=\"#\" data-view=\"select-address\">&lt; Back</a>\n      </div>\n      ${(0,_AddressSearch__WEBPACK_IMPORTED_MODULE_5__[\"default\"])()} \n    </section>\n    \n    <section class=\"bash--delivery-view\" data-section=\"address-form\">\n       <div class=\"bash--heading\">\n        <h3>Delivery address</h3>\n        <a href=\"#\" class=\"back-button--search\" data-view=\"address-search\">&lt; Back</a>\n        <a href=\"#\" class=\"back-button--select\" data-view=\"select-address\">&lt; Back</a>\n      </div>\n      ${(0,_AddressForm__WEBPACK_IMPORTED_MODULE_4__[\"default\"])()}\n    </section>\n    \n  </div>`;\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (DeliverContainer);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/DeliverContainer.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _utils_const__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/const */ \"./src/utils/const.js\");\n/* harmony import */ var _Addresses__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./Addresses */ \"./src/partials/Deliver/Addresses.js\");\n/* harmony import */ var _AddressForm__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./AddressForm */ \"./src/partials/Deliver/AddressForm.js\");\n/* harmony import */ var _AddressSearch__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./AddressSearch */ \"./src/partials/Deliver/AddressSearch.js\");\n/* harmony import */ var _DeliveryOptions__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./DeliveryOptions */ \"./src/partials/Deliver/DeliveryOptions.js\");\n/* harmony import */ var _MixedProducts__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./MixedProducts */ \"./src/partials/Deliver/MixedProducts.js\");\n/* harmony import */ var _TVorRICAMsg__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./TVorRICAMsg */ \"./src/partials/Deliver/TVorRICAMsg.js\");\n\n\n\n\n\n\n\n\nconst DeliverContainer = ({\n  hasFurn\n}) => `\n<div id=\"bash--shipping-messages\">\n${(0,_TVorRICAMsg__WEBPACK_IMPORTED_MODULE_6__[\"default\"])()}\n${(0,_MixedProducts__WEBPACK_IMPORTED_MODULE_5__[\"default\"])()}\n</div>\n  <div class=\"bash--delivery-container\" id=\"bash--delivery-container\" data-view=\"select-address\">\n  \n   <form id=\"bash--delivery-form\" name=\"bash--delivery-form\" method=\"post\">\n\n    <section class=\"bash--delivery-view\" data-section=\"select-address\">\n    <div class=\"bash--heading\">\n        <h2>Delivery address</h2>\n        <a href=\"#\" data-view=\"address-search\">Add address</a>\n      </div>\n      ${(0,_Addresses__WEBPACK_IMPORTED_MODULE_1__[\"default\"])()}\n    </section>\n\n    <section id=\"bash-delivery-options\" class=\"shipping-method bash--delivery-view\" data-section=\"select-address\">\n      <hr>\n      <div class=\"bash--heading sub-heading\">\n        <h3>Delivery method</h3>\n        ${hasFurn ? _utils_const__WEBPACK_IMPORTED_MODULE_0__.FURNITURE_FEE_LINK : ''}\n      </div>\n      ${(0,_DeliveryOptions__WEBPACK_IMPORTED_MODULE_4__[\"default\"])()}\n      <button \n        class=\"submit btn-go-to-payment btn btn-large btn-success\"\n        id=\"btn-save-delivery\" \n        type=\"submit\">\n          Go to payment\n      </button>\n    </section>\n   </form>\n\n    <section class=\"bash--delivery-view\" data-section=\"address-search\">\n      <div class=\"bash--heading\">\n        <h3>Add a new delivery address</h3>\n        <a href=\"#\" data-view=\"select-address\">&lt; Back</a>\n      </div>\n      ${(0,_AddressSearch__WEBPACK_IMPORTED_MODULE_3__[\"default\"])()} \n    </section>\n    \n    <section class=\"bash--delivery-view\" data-section=\"address-form\">\n       <div class=\"bash--heading\">\n        <h3>Delivery address</h3>\n        <a href=\"#\" class=\"back-button--search\" data-view=\"address-search\">&lt; Back</a>\n        <a href=\"#\" class=\"back-button--select\" data-view=\"select-address\">&lt; Back</a>\n      </div>\n      ${(0,_AddressForm__WEBPACK_IMPORTED_MODULE_2__[\"default\"])()}\n    </section>\n    \n  </div>`;\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (DeliverContainer);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/DeliverContainer.js?");
 
 /***/ }),
 
@@ -220,6 +220,16 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
+/***/ "./src/partials/Deliver/MixedProducts.js":
+/*!***********************************************!*\
+  !*** ./src/partials/Deliver/MixedProducts.js ***!
+  \***********************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\nconst MixedProducts = () => `\n  <div id=\"tfg-custom-mixed-msg\" class=\"tfg-custom-msg\">\n    <p class=\"tfg-custom-icon\"></p>\n    <p class=\"tfg-custom-text\">\n      We'll ship your furniture and other items in your cart to the selected address. \n      Only the furniture delivery fee will apply.\n      </p>\n  </div>\n`;\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (MixedProducts);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/MixedProducts.js?");
+
+/***/ }),
+
 /***/ "./src/partials/Deliver/RICAForm.js":
 /*!******************************************!*\
   !*** ./src/partials/Deliver/RICAForm.js ***!
@@ -240,6 +250,16 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
+/***/ "./src/partials/Deliver/TVorRICAMsg.js":
+/*!*********************************************!*\
+  !*** ./src/partials/Deliver/TVorRICAMsg.js ***!
+  \*********************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\nconst TVorRICAMsg = () => `\n  <div id=\"tfg-custom-tvrica-msg\" class=\"tfg-custom-msg\">\n    <p class=\"tfg-custom-icon\"></p>\n    <p class=\"tfg-custom-text\">\n      You can't collect this order in store because your cart contains items \n      which require either RICA or TV License validation.\n    </p>\n  </div>\n`;\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (TVorRICAMsg);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/TVorRICAMsg.js?");
+
+/***/ }),
+
 /***/ "./src/partials/Deliver/constants.js":
 /*!*******************************************!*\
   !*** ./src/partials/Deliver/constants.js ***!
@@ -256,7 +276,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \***************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"addressIsValid\": () => (/* binding */ addressIsValid),\n/* harmony export */   \"clearRicaFields\": () => (/* binding */ clearRicaFields),\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__),\n/* harmony export */   \"getBestRecipient\": () => (/* binding */ getBestRecipient),\n/* harmony export */   \"initGoogleAutocomplete\": () => (/* binding */ initGoogleAutocomplete),\n/* harmony export */   \"mapGoogleAddress\": () => (/* binding */ mapGoogleAddress),\n/* harmony export */   \"parseAttribute\": () => (/* binding */ parseAttribute),\n/* harmony export */   \"populateAddressForm\": () => (/* binding */ populateAddressForm),\n/* harmony export */   \"populateExtraFields\": () => (/* binding */ populateExtraFields),\n/* harmony export */   \"populateFurnitureFields\": () => (/* binding */ populateFurnitureFields),\n/* harmony export */   \"populateRicaFields\": () => (/* binding */ populateRicaFields),\n/* harmony export */   \"populateTVFields\": () => (/* binding */ populateTVFields),\n/* harmony export */   \"setAddress\": () => (/* binding */ setAddress),\n/* harmony export */   \"setCartClasses\": () => (/* binding */ setCartClasses),\n/* harmony export */   \"setDeliveryLoading\": () => (/* binding */ setDeliveryLoading),\n/* harmony export */   \"showHideLiftOrStairs\": () => (/* binding */ showHideLiftOrStairs),\n/* harmony export */   \"submitAddressForm\": () => (/* binding */ submitAddressForm),\n/* harmony export */   \"submitDeliveryForm\": () => (/* binding */ submitDeliveryForm),\n/* harmony export */   \"updateDeliveryFeeDisplay\": () => (/* binding */ updateDeliveryFeeDisplay)\n/* harmony export */ });\n/* harmony import */ var _utils_const__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/const */ \"./src/utils/const.js\");\n/* harmony import */ var _utils_functions__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/functions */ \"./src/utils/functions.js\");\n/* harmony import */ var _utils_services__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../utils/services */ \"./src/utils/services.js\");\n/* harmony import */ var _constants__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./constants */ \"./src/partials/Deliver/constants.js\");\n\n\n\n\nconst setDeliveryLoading = () => {\n  document.querySelector('.bash--delivery-container').classList.add('shimmer');\n};\nconst mapGoogleAddress = (addressComponents, geometry) => {\n  const streetNumber = addressComponents.find(item => item.types.includes('street_number'))?.long_name;\n  const street = addressComponents.find(item => item.types.includes('route'))?.long_name;\n  const neighborhood = addressComponents.find(item => item.types.includes('sublocality'))?.long_name;\n  const city = addressComponents.find(item => item.types.includes('locality'))?.long_name;\n  const postalCode = addressComponents.find(item => item.types.includes('postal_code'))?.long_name;\n  const state = addressComponents.find(item => item.types.includes('administrative_area_level_1'))?.long_name;\n  const coords = {\n    lat: '',\n    lng: ''\n  };\n\n  if (geometry) {\n    coords.lat = geometry.location.lat();\n    coords.lng = geometry.location.lng();\n  }\n\n  return {\n    street: `${streetNumber ?? ''} ${street ?? ''}`.trim(),\n    neighborhood,\n    city,\n    postalCode,\n    state,\n    ...coords\n  };\n};\n\nconst provinceShortCode = province => {\n  switch (province) {\n    case 'Select':\n      return '';\n\n    case 'Western Cape':\n      return 'WC';\n\n    case 'Easter Cape':\n      return 'EC';\n\n    case 'Gauteng':\n      return 'GP';\n\n    case 'KwaZulu-Natal':\n    case 'KwaZulu Natal':\n      return 'KZN';\n\n    case 'Northern Cape':\n      return 'NC';\n\n    case 'Limpopo':\n      return 'LP';\n\n    case 'Mpumalanga':\n      return 'MP';\n\n    case 'North West':\n      return 'NW';\n\n    case 'Freestate':\n    case 'Free State':\n      return 'FS';\n\n    default:\n      return province;\n  }\n};\n\nconst getBestRecipient = () => {\n  const receiverName = window?.vtexjs?.checkout?.orderForm?.shippingData?.address?.receiverName;\n  const firstName = window?.vtexjs?.checkout?.orderForm?.clientProfileData?.firstName;\n  const lastName = window?.vtexjs?.checkout?.orderForm?.clientProfileData?.lastName;\n  const clientProfileName = `${firstName ?? ''} ${lastName ?? ''}`.trim();\n  return receiverName || document.getElementById('client-first-name')?.value || clientProfileName;\n};\n\nconst populateAddressFromSearch = address => {\n  const {\n    street,\n    neighborhood,\n    postalCode,\n    state,\n    city\n  } = address; // Clear any populated fields\n\n  document.getElementById('bash--address-form').reset();\n  document.getElementById('bash--input-number').value = '';\n  document.getElementById('bash--input-street').value = street;\n  document.getElementById('bash--input-neighborhood').value = neighborhood;\n  document.getElementById('bash--input-city').value = city;\n  document.getElementById('bash--input-postalCode').value = postalCode;\n  document.getElementById('bash--input-state').value = provinceShortCode(state);\n};\n\nconst populateAddressForm = address => {\n  const {\n    street,\n    neighborhood,\n    postalCode,\n    state,\n    city,\n    receiverName,\n    complement,\n    id,\n    addressId,\n    addressName\n  } = address; // Clear any populated fields\n\n  document.getElementById('bash--address-form').reset(); // Only overwrite defaults if values exist.\n\n  if (receiverName) document.getElementById('bash--input-receiverName').value = receiverName;\n  if (complement) document.getElementById('bash--input-complement').value = complement; // addressId indicates that address is being edited / completed.\n\n  if (id || addressId) document.getElementById('bash--input-addressId').value = id || addressId; // TODO remove this?\n\n  if (addressName) document.getElementById('bash--input-addressName').value = addressName;\n  document.getElementById('bash--input-number').value = '';\n  document.getElementById('bash--input-street').value = street;\n  document.getElementById('bash--input-neighborhood').value = neighborhood;\n  document.getElementById('bash--input-city').value = city;\n  document.getElementById('bash--input-postalCode').value = postalCode;\n  document.getElementById('bash--input-state').value = provinceShortCode(state);\n};\nconst initGoogleAutocomplete = () => {\n  const input = document.getElementById('bash--input-address-search');\n  const autocomplete = new window.google.maps.places.Autocomplete(input, {\n    componentRestrictions: {\n      country: 'ZA'\n    }\n  });\n  window.google.maps.event.addListener(autocomplete, 'place_changed', () => {\n    const place = autocomplete.getPlace();\n    const {\n      address_components: addressComponents,\n      geometry\n    } = place;\n    const address = mapGoogleAddress(addressComponents, geometry); // Populate the form\n    // Set view to add-address\n\n    populateAddressFromSearch(address);\n    window.postMessage({\n      action: 'setDeliveryView',\n      view: 'address-form'\n    });\n  });\n};\nconst parseAttribute = data => JSON.parse(decodeURIComponent(data));\nconst populateExtraFields = (address, fields, prefix = '', override = false) => {\n  for (let i = 0; i < fields.length; i++) {\n    const fieldId = `bash--input-${prefix}${fields[i]}`;\n\n    if (document.getElementById(fieldId) && (address[fields[i]] || override) && (!document.getElementById(fieldId).value || override)) {\n      document.getElementById(fieldId).value = address[fields[i]];\n    }\n  }\n};\nconst populateRicaFields = async () => {\n  const {\n    address\n  } = window.vtexjs.checkout.orderForm.shippingData;\n  address.fullName = getBestRecipient();\n  address.streetAddress = address.street;\n  address.suburb = address.neighborhood;\n  address.province = address.state;\n  populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields, 'rica_');\n  const data = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP);\n  console.info('populateRicaFields', {\n    data\n  });\n  if (data.streetAddress) populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields, 'rica_', true);\n};\nconst clearRicaFields = () => {\n  const clearedRica = {\n    fullName: '',\n    streetAddress: '',\n    suburb: '',\n    city: '',\n    postalCode: '',\n    province: ''\n  };\n  populateExtraFields(clearedRica, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields, 'rica_', true);\n};\nconst showHideLiftOrStairs = floor => {\n  if (floor && floor !== 'ground') {\n    $('.bash--dropdownfield-liftOrStairs').slideDown().addClass('required');\n    $('#bash--input-liftOrStairs').attr('required', 'required');\n  } else {\n    $('.bash--dropdownfield-liftOrStairs').slideUp();\n    $('#bash--input-liftOrStairs').removeAttr('required');\n  }\n};\nconst populateFurnitureFields = async () => {\n  const data = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.FURNITURE_APP);\n  console.info('populateFurnitureFields', {\n    data\n  });\n  populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields);\n  showHideLiftOrStairs(data?.deliveryFloor);\n};\nconst populateTVFields = async () => {\n  const data = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP);\n  console.info('populateTVFields', {\n    data\n  });\n  populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields, 'tv');\n}; // Runs when you setAddress\n\nconst addressIsValid = (address, validateExtraFields = true) => {\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items);\n  let requiredFields = [];\n  const invalidFields = [];\n  requiredFields = [..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredAddressFields];\n\n  if (hasFurniture && validateExtraFields) {\n    requiredFields = [...requiredFields, ..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields];\n  }\n\n  if (hasTVs && validateExtraFields) {\n    requiredFields = [...requiredFields, ..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields];\n  }\n\n  if (hasSimCards && validateExtraFields) {\n    requiredFields = [...requiredFields, ..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields];\n  }\n\n  for (let i = 0; i < requiredFields.length; i++) {\n    if (!address[requiredFields[i]]) invalidFields.push(requiredFields[i]);\n  }\n\n  return {\n    isValid: !invalidFields.length,\n    invalidFields\n  };\n};\nconst setCartClasses = () => {\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards,\n    hasFurnitureMixed\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items);\n  const $container = '#shipping-data';\n\n  if (hasFurniture) {\n    $(`${$container}:not(.has-furniture)`).addClass('has-furniture');\n  } else {\n    $(`${$container}.has-furniture`).removeClass('has-furniture');\n  }\n\n  if (hasTVs) {\n    $(`${$container}:not(.has-tv)`).addClass('has-tv');\n  } else {\n    $(`${$container}.has-tv`).removeClass('has-tv');\n  }\n\n  if (hasSimCards) {\n    $(`${$container}:not(.has-rica)`).addClass('has-rica');\n  } else {\n    $(`${$container}.has-rica`).removeClass('has-rica');\n  }\n\n  if (hasFurnitureMixed) {\n    $(`${$container}:not(.has-furniture-mixed)`).addClass('has-furniture-mixed');\n  } else {\n    $(`${$container}.has-furniture-mixed`).removeClass('has-furniture-mixed');\n  }\n};\nconst updateDeliveryFeeDisplay = () => {\n  if (!window.vtexjs) return;\n  const {\n    value: shippingFee\n  } = window.vtexjs.checkout.orderForm.totalizers.find(item => item.id === 'Shipping');\n  let feeText = 'Free';\n  if (shippingFee > 0) feeText = `R${(shippingFee / 100).toFixed(2).replace('.00', '')}`;\n\n  if ($('#bash--delivery-fee').length > 0) {\n    document.getElementById('bash--delivery-fee').innerHTML = feeText;\n  }\n}; // TODO move somewhere else?\n\nconst setAddress = (address, options = {\n  validateExtraFields: true\n}) => {\n  const {\n    validateExtraFields\n  } = options;\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items);\n\n  if (hasFurniture) {\n    populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields);\n    showHideLiftOrStairs(address.deliveryFloor);\n  }\n\n  if (hasTVs) populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields, 'tv_');\n  if (hasSimCards) populateRicaFields();\n  const {\n    isValid,\n    invalidFields\n  } = addressIsValid(address, validateExtraFields);\n\n  if (!isValid) {\n    populateAddressForm(address);\n    $('#bash--address-form').addClass('show-form-errors');\n    if (validateExtraFields) $('#bash--delivery-form')?.addClass('show-form-errors');\n    $(`#bash--input-${invalidFields[0]}`).focus();\n\n    if (_constants__WEBPACK_IMPORTED_MODULE_3__.requiredAddressFields.includes(invalidFields[0])) {\n      window.postMessage({\n        action: 'setDeliveryView',\n        view: 'address-form'\n      });\n    }\n\n    return;\n  } // Fix bad addressType.\n\n\n  if (address.addressType === 'business') address.addressType = 'commercial';\n  if (!_constants__WEBPACK_IMPORTED_MODULE_3__.validAddressTypes.includes(address.addressType)) address.addressType = 'residential';\n  const {\n    shippingData\n  } = window?.vtexjs?.checkout?.orderForm;\n  shippingData.address = address;\n  shippingData.address.number = shippingData.address.number ?? ' ';\n  shippingData.selectedAddresses = [address]; // Start Shimmering\n\n  setDeliveryLoading();\n  window.vtexjs.checkout.sendAttachment('shippingData', shippingData).then(() => {\n    // End shimmer\n    if (address.addressName) {\n      (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.updateAddressListing)(shippingData.address);\n    }\n  }).done(() => (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.clearLoaders)());\n};\nconst submitAddressForm = async event => {\n  event.preventDefault(); // Prevent false positive for invalid selects.\n\n  $('select').change();\n  const form = document.forms['bash--address-form'];\n  const addressName = $('#bash--input-addressName').val();\n  const storedAddress = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getAddressByName)(addressName);\n  const fields = ['addressId', 'addressName', 'addressType', 'receiverName', 'postalCode', 'city', 'state', 'country', 'street', 'neighborhood', 'complement'];\n  const address = {\n    isDisposable: false,\n    reference: null,\n    geoCoordinates: [],\n    number: '',\n    country: 'ZAF',\n    ...storedAddress\n  };\n\n  for (let f = 0; f < fields.length; f++) {\n    address[fields[f]] = form[fields[f]]?.value || null;\n  }\n\n  address.addressName = address.addressName || address.addressId;\n  address.addressId = address.addressId || address.addressName;\n  const {\n    isValid,\n    invalidFields\n  } = addressIsValid(address, false);\n\n  if (!isValid) {\n    $('#bash--address-form').addClass('show-form-errors');\n    $(`#bash--input-${invalidFields[0]}`).focus();\n\n    if (_constants__WEBPACK_IMPORTED_MODULE_3__.requiredAddressFields.includes(invalidFields[0])) {\n      window.postMessage({\n        action: 'setDeliveryView',\n        view: 'address-form'\n      });\n    }\n\n    return;\n  } // Apply the selected address to customers orderForm.\n\n\n  await setAddress(address, {\n    validateExtraFields: false\n  }); // Update the localstore, and the API\n\n  await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.addOrUpdateAddress)(address);\n  window.postMessage({\n    action: 'setDeliveryView',\n    view: 'select-address'\n  });\n};\nconst submitDeliveryForm = async event => {\n  event.preventDefault();\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    address\n  } = window.vtexjs.checkout.orderForm.shippingData;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items); // Prevent false positive for invalid selects.\n\n  $('select').change();\n  let fullAddress = {};\n  setDeliveryLoading();\n  const dbAddress = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getAddressByName)($(\"[name='selected-address']:checked\").val());\n  fullAddress = { ...address,\n    ...dbAddress\n  };\n  const furnitureData = {};\n  const ricaData = {};\n  const tvData = {};\n\n  if (hasFurniture) {\n    const fields = _constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields;\n\n    for (let i = 0; i < fields.length; i++) {\n      if (!address[fields[i]]) fullAddress[fields[i]] = $(`#bash--input-${fields[i]}`).val();\n      furnitureData[fields[i]] = $(`#bash--input-${fields[i]}`).val();\n    }\n\n    const furnitureDataSent = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.sendOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.FURNITURE_APP, furnitureData);\n    console.info({\n      furnitureDataSent\n    });\n  } // Not saved to address profile.\n\n\n  if (hasSimCards) {\n    const fields = _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields;\n\n    for (let i = 0; i < fields.length; i++) {\n      ricaData[fields[i]] = $(`#bash--input-rica_${fields[i]}`).val();\n    }\n\n    const ricaDataSent = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.sendOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP, ricaData);\n    console.info({\n      ricaDataSent\n    });\n  }\n\n  if (hasTVs) {\n    const fields = _constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields;\n\n    for (let i = 0; i < fields.length; i++) {\n      if (!address[fields[i]]) fullAddress[fields[i]] = $(`#bash--input-tv_${fields[i]}`).val();\n      tvData[fields[i]] = $(`#bash--input-tv_${fields[i]}`).val();\n    }\n\n    const tvDataSent = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.sendOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP, tvData);\n    console.info({\n      tvDataSent\n    });\n  }\n\n  await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.addOrUpdateAddress)(fullAddress);\n  window.location.hash = _utils_const__WEBPACK_IMPORTED_MODULE_0__.STEPS.PAYMENT;\n  (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.clearLoaders)();\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (mapGoogleAddress);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/utils.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"addressIsValid\": () => (/* binding */ addressIsValid),\n/* harmony export */   \"clearRicaFields\": () => (/* binding */ clearRicaFields),\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__),\n/* harmony export */   \"getBestRecipient\": () => (/* binding */ getBestRecipient),\n/* harmony export */   \"initGoogleAutocomplete\": () => (/* binding */ initGoogleAutocomplete),\n/* harmony export */   \"mapGoogleAddress\": () => (/* binding */ mapGoogleAddress),\n/* harmony export */   \"parseAttribute\": () => (/* binding */ parseAttribute),\n/* harmony export */   \"populateAddressForm\": () => (/* binding */ populateAddressForm),\n/* harmony export */   \"populateExtraFields\": () => (/* binding */ populateExtraFields),\n/* harmony export */   \"populateFurnitureFields\": () => (/* binding */ populateFurnitureFields),\n/* harmony export */   \"populateRicaFields\": () => (/* binding */ populateRicaFields),\n/* harmony export */   \"populateTVFields\": () => (/* binding */ populateTVFields),\n/* harmony export */   \"setAddress\": () => (/* binding */ setAddress),\n/* harmony export */   \"setCartClasses\": () => (/* binding */ setCartClasses),\n/* harmony export */   \"setDeliveryLoading\": () => (/* binding */ setDeliveryLoading),\n/* harmony export */   \"showHideLiftOrStairs\": () => (/* binding */ showHideLiftOrStairs),\n/* harmony export */   \"submitAddressForm\": () => (/* binding */ submitAddressForm),\n/* harmony export */   \"submitDeliveryForm\": () => (/* binding */ submitDeliveryForm),\n/* harmony export */   \"updateDeliveryFeeDisplay\": () => (/* binding */ updateDeliveryFeeDisplay)\n/* harmony export */ });\n/* harmony import */ var _utils_const__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/const */ \"./src/utils/const.js\");\n/* harmony import */ var _utils_functions__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/functions */ \"./src/utils/functions.js\");\n/* harmony import */ var _utils_services__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../utils/services */ \"./src/utils/services.js\");\n/* harmony import */ var _constants__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./constants */ \"./src/partials/Deliver/constants.js\");\n\n\n\n\nconst setDeliveryLoading = () => {\n  document.querySelector('.bash--delivery-container').classList.add('shimmer');\n};\nconst mapGoogleAddress = (addressComponents, geometry) => {\n  const streetNumber = addressComponents.find(item => item.types.includes('street_number'))?.long_name;\n  const street = addressComponents.find(item => item.types.includes('route'))?.long_name;\n  const neighborhood = addressComponents.find(item => item.types.includes('sublocality'))?.long_name;\n  const city = addressComponents.find(item => item.types.includes('locality'))?.long_name;\n  const postalCode = addressComponents.find(item => item.types.includes('postal_code'))?.long_name;\n  const state = addressComponents.find(item => item.types.includes('administrative_area_level_1'))?.long_name;\n  const coords = {\n    lat: '',\n    lng: ''\n  };\n\n  if (geometry) {\n    coords.lat = geometry.location.lat();\n    coords.lng = geometry.location.lng();\n  }\n\n  return {\n    street: `${streetNumber ?? ''} ${street ?? ''}`.trim(),\n    neighborhood,\n    city,\n    postalCode,\n    state,\n    ...coords\n  };\n};\n\nconst provinceShortCode = province => {\n  switch (province) {\n    case 'Select':\n      return '';\n\n    case 'Western Cape':\n      return 'WC';\n\n    case 'Easter Cape':\n      return 'EC';\n\n    case 'Gauteng':\n      return 'GP';\n\n    case 'KwaZulu-Natal':\n    case 'KwaZulu Natal':\n      return 'KZN';\n\n    case 'Northern Cape':\n      return 'NC';\n\n    case 'Limpopo':\n      return 'LP';\n\n    case 'Mpumalanga':\n      return 'MP';\n\n    case 'North West':\n      return 'NW';\n\n    case 'Freestate':\n    case 'Free State':\n      return 'FS';\n\n    default:\n      return province;\n  }\n};\n\nconst getBestRecipient = () => {\n  const receiverName = window?.vtexjs?.checkout?.orderForm?.shippingData?.address?.receiverName;\n  const firstName = window?.vtexjs?.checkout?.orderForm?.clientProfileData?.firstName;\n  const lastName = window?.vtexjs?.checkout?.orderForm?.clientProfileData?.lastName;\n  const clientProfileName = `${firstName ?? ''} ${lastName ?? ''}`.trim();\n  return receiverName || document.getElementById('client-first-name')?.value || clientProfileName;\n};\n\nconst populateAddressFromSearch = address => {\n  const {\n    street,\n    neighborhood,\n    postalCode,\n    state,\n    city\n  } = address; // Clear any populated fields\n\n  document.getElementById('bash--address-form').reset();\n  document.getElementById('bash--input-number').value = '';\n  document.getElementById('bash--input-street').value = street;\n  document.getElementById('bash--input-neighborhood').value = neighborhood;\n  document.getElementById('bash--input-city').value = city;\n  document.getElementById('bash--input-postalCode').value = postalCode;\n  document.getElementById('bash--input-state').value = provinceShortCode(state);\n};\n\nconst populateAddressForm = address => {\n  const {\n    street,\n    neighborhood,\n    postalCode,\n    state,\n    city,\n    receiverName,\n    complement,\n    id,\n    addressId,\n    addressName\n  } = address; // Clear any populated fields\n\n  document.getElementById('bash--address-form').reset(); // Only overwrite defaults if values exist.\n\n  if (receiverName) document.getElementById('bash--input-receiverName').value = receiverName;\n  if (complement) document.getElementById('bash--input-complement').value = complement; // addressId indicates that address is being edited / completed.\n\n  if (id || addressId) document.getElementById('bash--input-addressId').value = id || addressId; // TODO remove this?\n\n  if (addressName) document.getElementById('bash--input-addressName').value = addressName;\n  document.getElementById('bash--input-number').value = '';\n  document.getElementById('bash--input-street').value = street;\n  document.getElementById('bash--input-neighborhood').value = neighborhood;\n  document.getElementById('bash--input-city').value = city;\n  document.getElementById('bash--input-postalCode').value = postalCode;\n  document.getElementById('bash--input-state').value = provinceShortCode(state);\n};\nconst initGoogleAutocomplete = () => {\n  const input = document.getElementById('bash--input-address-search');\n  const autocomplete = new window.google.maps.places.Autocomplete(input, {\n    componentRestrictions: {\n      country: 'ZA'\n    }\n  });\n  window.google.maps.event.addListener(autocomplete, 'place_changed', () => {\n    const place = autocomplete.getPlace();\n    const {\n      address_components: addressComponents,\n      geometry\n    } = place;\n    const address = mapGoogleAddress(addressComponents, geometry); // Populate the form\n    // Set view to add-address\n\n    populateAddressFromSearch(address);\n    window.postMessage({\n      action: 'setDeliveryView',\n      view: 'address-form'\n    });\n  });\n};\nconst parseAttribute = data => JSON.parse(decodeURIComponent(data));\nconst populateExtraFields = (address, fields, prefix = '', override = false) => {\n  for (let i = 0; i < fields.length; i++) {\n    const fieldId = `bash--input-${prefix}${fields[i]}`;\n\n    if (document.getElementById(fieldId) && (address[fields[i]] || override) && (!document.getElementById(fieldId).value || override)) {\n      document.getElementById(fieldId).value = address[fields[i]];\n    }\n  }\n};\nconst populateRicaFields = async () => {\n  const {\n    address\n  } = window.vtexjs.checkout.orderForm.shippingData;\n  address.fullName = getBestRecipient();\n  address.streetAddress = address.street;\n  address.suburb = address.neighborhood;\n  address.province = address.state;\n  populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields, 'rica_');\n  const data = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP);\n  console.info('populateRicaFields', {\n    data\n  });\n  if (data.streetAddress) populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields, 'rica_', true);\n};\nconst clearRicaFields = () => {\n  const clearedRica = {\n    idOrPassport: '',\n    // TODO populate with users ID\n    fullName: '',\n    streetAddress: '',\n    suburb: '',\n    city: '',\n    postalCode: '',\n    province: ''\n  };\n  populateExtraFields(clearedRica, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields, 'rica_', true);\n};\nconst showHideLiftOrStairs = floor => {\n  if (floor && floor !== 'ground') {\n    $('.bash--dropdownfield-liftOrStairs').slideDown().addClass('required');\n    $('#bash--input-liftOrStairs').attr('required', 'required');\n  } else {\n    $('.bash--dropdownfield-liftOrStairs').slideUp();\n    $('#bash--input-liftOrStairs').removeAttr('required');\n  }\n};\nconst populateFurnitureFields = async () => {\n  const data = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.FURNITURE_APP);\n  console.info('populateFurnitureFields', {\n    data\n  });\n  populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields);\n  showHideLiftOrStairs(data?.deliveryFloor);\n};\nconst populateTVFields = async () => {\n  const data = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP);\n  console.info('populateTVFields', {\n    data\n  });\n  populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields, 'tv');\n}; // Runs when you setAddress\n\nconst addressIsValid = (address, validateExtraFields = true) => {\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items);\n  let requiredFields = [];\n  const invalidFields = [];\n  requiredFields = [..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredAddressFields];\n\n  if (hasFurniture && validateExtraFields) {\n    requiredFields = [...requiredFields, ..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields];\n  }\n\n  if (hasTVs && validateExtraFields) {\n    requiredFields = [...requiredFields, ..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields];\n  }\n\n  if (hasSimCards && validateExtraFields) {\n    requiredFields = [...requiredFields, ..._constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields];\n  }\n\n  for (let i = 0; i < requiredFields.length; i++) {\n    if (!address[requiredFields[i]]) invalidFields.push(requiredFields[i]);\n  }\n\n  return {\n    isValid: !invalidFields.length,\n    invalidFields\n  };\n};\nconst setCartClasses = () => {\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards,\n    hasFurnitureMixed\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items);\n  const $container = '#shipping-data';\n\n  if (hasFurniture) {\n    $(`${$container}:not(.has-furniture)`).addClass('has-furniture');\n  } else {\n    $(`${$container}.has-furniture`).removeClass('has-furniture');\n  }\n\n  if (hasTVs) {\n    $(`${$container}:not(.has-tv)`).addClass('has-tv');\n  } else {\n    $(`${$container}.has-tv`).removeClass('has-tv');\n  }\n\n  if (hasSimCards) {\n    $(`${$container}:not(.has-rica)`).addClass('has-rica');\n  } else {\n    $(`${$container}.has-rica`).removeClass('has-rica');\n  }\n\n  if (hasFurnitureMixed) {\n    $(`${$container}:not(.has-furniture-mixed)`).addClass('has-furniture-mixed');\n  } else {\n    $(`${$container}.has-furniture-mixed`).removeClass('has-furniture-mixed');\n  }\n};\nconst updateDeliveryFeeDisplay = () => {\n  if (!window.vtexjs) return;\n  const {\n    value: shippingFee\n  } = window.vtexjs.checkout.orderForm.totalizers.find(item => item.id === 'Shipping');\n  let feeText = 'Free';\n  if (shippingFee > 0) feeText = `R${(shippingFee / 100).toFixed(2).replace('.00', '')}`;\n\n  if ($('#bash--delivery-fee').length > 0) {\n    document.getElementById('bash--delivery-fee').innerHTML = feeText;\n  }\n}; // TODO move somewhere else?\n\nconst setAddress = (address, options = {\n  validateExtraFields: true\n}) => {\n  const {\n    validateExtraFields\n  } = options;\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items);\n\n  if (hasFurniture) {\n    populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields);\n    showHideLiftOrStairs(address.deliveryFloor);\n  }\n\n  if (hasTVs) populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields, 'tv_');\n  if (hasSimCards) populateRicaFields();\n  const {\n    isValid,\n    invalidFields\n  } = addressIsValid(address, validateExtraFields);\n\n  if (!isValid) {\n    populateAddressForm(address);\n    $('#bash--address-form').addClass('show-form-errors');\n    if (validateExtraFields) $('#bash--delivery-form')?.addClass('show-form-errors');\n    $(`#bash--input-${invalidFields[0]}`).focus();\n\n    if (_constants__WEBPACK_IMPORTED_MODULE_3__.requiredAddressFields.includes(invalidFields[0])) {\n      window.postMessage({\n        action: 'setDeliveryView',\n        view: 'address-form'\n      });\n    }\n\n    return;\n  } // Fix bad addressType.\n\n\n  if (address.addressType === 'business') address.addressType = 'commercial';\n  if (!_constants__WEBPACK_IMPORTED_MODULE_3__.validAddressTypes.includes(address.addressType)) address.addressType = 'residential';\n  const {\n    shippingData\n  } = window?.vtexjs?.checkout?.orderForm;\n  shippingData.address = address;\n  shippingData.address.number = shippingData.address.number ?? ' ';\n  shippingData.selectedAddresses = [address]; // Start Shimmering\n\n  setDeliveryLoading();\n  window.vtexjs.checkout.sendAttachment('shippingData', shippingData).then(() => {\n    // End shimmer\n    if (address.addressName) {\n      (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.updateAddressListing)(shippingData.address);\n    }\n  }).done(() => (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.clearLoaders)());\n};\nconst submitAddressForm = async event => {\n  event.preventDefault(); // Prevent false positive for invalid selects.\n\n  $('select').change();\n  const form = document.forms['bash--address-form'];\n  const addressName = $('#bash--input-addressName').val();\n  const storedAddress = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getAddressByName)(addressName);\n  const fields = ['addressId', 'addressName', 'addressType', 'receiverName', 'postalCode', 'city', 'state', 'country', 'street', 'neighborhood', 'complement'];\n  const address = {\n    isDisposable: false,\n    reference: null,\n    geoCoordinates: [],\n    number: '',\n    country: 'ZAF',\n    ...storedAddress\n  };\n\n  for (let f = 0; f < fields.length; f++) {\n    address[fields[f]] = form[fields[f]]?.value || null;\n  }\n\n  address.addressName = address.addressName || address.addressId;\n  address.addressId = address.addressId || address.addressName;\n  const {\n    isValid,\n    invalidFields\n  } = addressIsValid(address, false);\n\n  if (!isValid) {\n    $('#bash--address-form').addClass('show-form-errors');\n    $(`#bash--input-${invalidFields[0]}`).focus();\n\n    if (_constants__WEBPACK_IMPORTED_MODULE_3__.requiredAddressFields.includes(invalidFields[0])) {\n      window.postMessage({\n        action: 'setDeliveryView',\n        view: 'address-form'\n      });\n    }\n\n    return;\n  } // Apply the selected address to customers orderForm.\n\n\n  await setAddress(address, {\n    validateExtraFields: false\n  }); // Update the localstore, and the API\n\n  await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.addOrUpdateAddress)(address);\n  window.postMessage({\n    action: 'setDeliveryView',\n    view: 'select-address'\n  });\n};\nconst submitDeliveryForm = async event => {\n  event.preventDefault();\n  const {\n    items\n  } = window.vtexjs.checkout.orderForm;\n  const {\n    address\n  } = window.vtexjs.checkout.orderForm.shippingData;\n  const {\n    hasFurniture,\n    hasTVs,\n    hasSimCards\n  } = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items); // Prevent false positive for invalid selects.\n\n  $('select').change();\n  let fullAddress = {};\n  setDeliveryLoading();\n  const dbAddress = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.getAddressByName)($(\"[name='selected-address']:checked\").val());\n  fullAddress = { ...address,\n    ...dbAddress\n  };\n  const furnitureData = {};\n  const ricaData = {};\n  const tvData = {};\n\n  if (hasFurniture) {\n    const fields = _constants__WEBPACK_IMPORTED_MODULE_3__.requiredFurnitureFields;\n\n    for (let i = 0; i < fields.length; i++) {\n      if (!address[fields[i]]) fullAddress[fields[i]] = $(`#bash--input-${fields[i]}`).val();\n      furnitureData[fields[i]] = $(`#bash--input-${fields[i]}`).val();\n    }\n\n    const furnitureDataSent = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.sendOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.FURNITURE_APP, furnitureData);\n    console.info({\n      furnitureDataSent\n    });\n  } // Not saved to address profile.\n\n\n  if (hasSimCards) {\n    const fields = _constants__WEBPACK_IMPORTED_MODULE_3__.requiredRicaFields;\n\n    for (let i = 0; i < fields.length; i++) {\n      ricaData[fields[i]] = $(`#bash--input-rica_${fields[i]}`).val();\n    }\n\n    const ricaDataSent = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.sendOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP, ricaData);\n    console.info({\n      ricaDataSent\n    });\n  }\n\n  if (hasTVs) {\n    const fields = _constants__WEBPACK_IMPORTED_MODULE_3__.requiredTVFields;\n\n    for (let i = 0; i < fields.length; i++) {\n      if (!address[fields[i]]) fullAddress[fields[i]] = $(`#bash--input-tv_${fields[i]}`).val();\n      tvData[fields[i]] = $(`#bash--input-tv_${fields[i]}`).val();\n    }\n\n    const tvDataSent = await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.sendOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP, tvData);\n    console.info({\n      tvDataSent\n    });\n  }\n\n  await (0,_utils_services__WEBPACK_IMPORTED_MODULE_2__.addOrUpdateAddress)(fullAddress);\n  window.location.hash = _utils_const__WEBPACK_IMPORTED_MODULE_0__.STEPS.PAYMENT;\n  (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.clearLoaders)();\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (mapGoogleAddress);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/utils.js?");
 
 /***/ }),
 
@@ -280,16 +300,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
-/***/ "./src/partials/MixedProducts.js":
-/*!***************************************!*\
-  !*** ./src/partials/MixedProducts.js ***!
-  \***************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\nconst MixedProducts = () => `\n  <div id=\"tfg-custom-mixed-msg\" class=\"tfg-custom-msg\">\n    <p class=\"tfg-custom-icon\"></p>\n    <p class=\"tfg-custom-text\">\n      We'll ship your furniture and other items in your cart to the selected address. \n      Only the furniture delivery fee will apply.\n      </p>\n  </div>\n`;\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (MixedProducts);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/MixedProducts.js?");
-
-/***/ }),
-
 /***/ "./src/partials/RICAForm.js":
 /*!**********************************!*\
   !*** ./src/partials/RICAForm.js ***!
@@ -310,23 +320,13 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
-/***/ "./src/partials/TVorRICAMsg.js":
-/*!*************************************!*\
-  !*** ./src/partials/TVorRICAMsg.js ***!
-  \*************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\nconst TVorRICAMsg = () => `\n  <div id=\"tfg-custom-tvrica-msg\" class=\"tfg-custom-msg\">\n    <p class=\"tfg-custom-icon\"></p>\n    <p class=\"tfg-custom-text\">\n      You can't collect this order in store because your cart contains items \n      which require either RICA or TV License validation.\n    </p>\n  </div>\n`;\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (TVorRICAMsg);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/TVorRICAMsg.js?");
-
-/***/ }),
-
 /***/ "./src/partials/index.js":
 /*!*******************************!*\
   !*** ./src/partials/index.js ***!
   \*******************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"AddressForm\": () => (/* reexport safe */ _AddressForm__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"AlertBox\": () => (/* reexport safe */ _AlertBox__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"FurnitureForm\": () => (/* reexport safe */ _FurnitureForm__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"InputError\": () => (/* reexport safe */ _InputError__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"MixedProducts\": () => (/* reexport safe */ _MixedProducts__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"PickupComplementField\": () => (/* reexport safe */ _AddressForm__WEBPACK_IMPORTED_MODULE_6__.PickupComplementField),\n/* harmony export */   \"RICAForm\": () => (/* reexport safe */ _RICAForm__WEBPACK_IMPORTED_MODULE_3__[\"default\"]),\n/* harmony export */   \"SuburbField\": () => (/* reexport safe */ _AddressForm__WEBPACK_IMPORTED_MODULE_6__.SuburbField),\n/* harmony export */   \"TVIDForm\": () => (/* reexport safe */ _TVIDForm__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"TVorRICAMsg\": () => (/* reexport safe */ _TVorRICAMsg__WEBPACK_IMPORTED_MODULE_1__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _FurnitureForm__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./FurnitureForm */ \"./src/partials/FurnitureForm.js\");\n/* harmony import */ var _TVorRICAMsg__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./TVorRICAMsg */ \"./src/partials/TVorRICAMsg.js\");\n/* harmony import */ var _TVIDForm__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./TVIDForm */ \"./src/partials/TVIDForm.js\");\n/* harmony import */ var _RICAForm__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./RICAForm */ \"./src/partials/RICAForm.js\");\n/* harmony import */ var _InputError__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./InputError */ \"./src/partials/InputError.js\");\n/* harmony import */ var _MixedProducts__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./MixedProducts */ \"./src/partials/MixedProducts.js\");\n/* harmony import */ var _AddressForm__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./AddressForm */ \"./src/partials/AddressForm.js\");\n/* harmony import */ var _AlertBox__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./AlertBox */ \"./src/partials/AlertBox.js\");\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"AddressForm\": () => (/* reexport safe */ _AddressForm__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"AlertBox\": () => (/* reexport safe */ _AlertBox__WEBPACK_IMPORTED_MODULE_1__[\"default\"]),\n/* harmony export */   \"FurnitureForm\": () => (/* reexport safe */ _FurnitureForm__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"InputError\": () => (/* reexport safe */ _InputError__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"MixedProducts\": () => (/* reexport safe */ _Deliver_MixedProducts__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"PickupComplementField\": () => (/* reexport safe */ _AddressForm__WEBPACK_IMPORTED_MODULE_0__.PickupComplementField),\n/* harmony export */   \"RICAForm\": () => (/* reexport safe */ _RICAForm__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"SuburbField\": () => (/* reexport safe */ _AddressForm__WEBPACK_IMPORTED_MODULE_0__.SuburbField),\n/* harmony export */   \"TVIDForm\": () => (/* reexport safe */ _TVIDForm__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"TVorRICAMsg\": () => (/* reexport safe */ _Deliver_TVorRICAMsg__WEBPACK_IMPORTED_MODULE_3__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _AddressForm__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./AddressForm */ \"./src/partials/AddressForm.js\");\n/* harmony import */ var _AlertBox__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./AlertBox */ \"./src/partials/AlertBox.js\");\n/* harmony import */ var _Deliver_MixedProducts__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Deliver/MixedProducts */ \"./src/partials/Deliver/MixedProducts.js\");\n/* harmony import */ var _Deliver_TVorRICAMsg__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./Deliver/TVorRICAMsg */ \"./src/partials/Deliver/TVorRICAMsg.js\");\n/* harmony import */ var _FurnitureForm__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./FurnitureForm */ \"./src/partials/FurnitureForm.js\");\n/* harmony import */ var _InputError__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./InputError */ \"./src/partials/InputError.js\");\n/* harmony import */ var _RICAForm__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./RICAForm */ \"./src/partials/RICAForm.js\");\n/* harmony import */ var _TVIDForm__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./TVIDForm */ \"./src/partials/TVIDForm.js\");\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/index.js?");
 
 /***/ }),
