+++ checkout-ui-custom/src/partials/Deliver/TVorRICAMsg.js
