+++ checkout-ui-custom/src/partials/index.js
@@ -1,20 +1,21 @@
-import FurnitureForm from './FurnitureForm';
-import TVorRICAMsg from './TVorRICAMsg';
-import TVIDForm from './TVIDForm';
-import RICAForm from './RICAForm';
-import InputError from './InputError';
-import MixedProducts from './MixedProducts';
-import AddressForm, { SuburbField, PickupComplementField } from './AddressForm';
+import AddressForm, { PickupComplementField, SuburbField } from './AddressForm';
 import AlertBox from './AlertBox';
+import MixedProducts from './Deliver/MixedProducts';
+import TVorRICAMsg from './Deliver/TVorRICAMsg';
+import FurnitureForm from './FurnitureForm';
+import InputError from './InputError';
+import RICAForm from './RICAForm';
+import TVIDForm from './TVIDForm';
 
 export {
   FurnitureForm,
   TVorRICAMsg,
-  TVIDForm, RICAForm,
+  TVIDForm,
+  RICAForm,
   InputError,
   MixedProducts,
   AddressForm,
   SuburbField,
   AlertBox,
-  PickupComplementField
+  PickupComplementField,
 };
